+++ machine/cpp/render/backend/pass/library.h
@@ -46,7 +46,7 @@ struct GxGpuPipelineState {
 	u32 displayModeWord = 0u;
 	u32 displayStartWord = 0u;
 	const std::array<u8, GX_GPU_VRAM_BYTE_COUNT>* vramSnapshotBytes = nullptr;
-	u32 vramSnapshotSerial = 0u;
+	u64 vramSnapshotSerial = 0u;
 };
 
 struct CRTPipelineOptions {
